Extract virtual pad skin model from settings store
